Update README.md
edited readme to reflect fixed GOV14 fields (also renamed these fields to fit naming conventions). Also removed this sentence: "Election data was compiled at the precinct level by a private individual and was found to be within 1% of the official state-reported results at the county and state levels." ...as data were not within 1% of county totals in the original shapefile for GOV14 (many missing R votes and some missing D votes in Beaver county, and flipped parties in Erie and Sullivan counties), and the remaining fields have not been double-checked.